Switch outcomes to floats

+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/repositories/burdenestimates/BurdenEstimateWriter.kt
@@ -133,7 +133,7 @@ abstract class BurdenEstimateWriter(
             estimate: BurdenEstimateWithRunId,
             countryId: Short,
             outcomeId: Short,
-            outcomeValue: BigDecimal?
+            outcomeValue: Float?
     ): List<Any?>
     {
         return listOf(
