+++ src/databaseTests/src/test/kotlin/org/vaccineimpact/api/databaseTests/tests/burdenEstimateRepository/BurdenEstimateRepositoryTests.kt
@@ -153,12 +153,12 @@ abstract class BurdenEstimateRepositoryTests : RepositoryTests<BurdenEstimateRep
     protected fun checkBurdenEstimates(db: JooqContext, setId: Int)
     {
         val records = getEstimatesInOrder(db)
-        checkRecord(records[0], setId, 2000, 50, "AFG", "cases", 100.toDecimal())
-        checkRecord(records[1], setId, 2000, 50, "AFG", "cohort_size", 1000.toDecimal())
-        checkRecord(records[2], setId, 2000, 50, "AFG", "deaths", 10.toDecimal())
-        checkRecord(records[3], setId, 1980, 30, "AGO", "cohort_size", 2000.toDecimal())
-        checkRecord(records[4], setId, 1980, 30, "AGO", "dalys", 73.6.toDecimal())
-        checkRecord(records[5], setId, 1980, 30, "AGO", "deaths", 20.toDecimal())
+        checkRecord(records[0], setId, 2000, 50, "AFG", "cases", 100F)
+        checkRecord(records[1], setId, 2000, 50, "AFG", "cohort_size", 1000F)
+        checkRecord(records[2], setId, 2000, 50, "AFG", "deaths", 10F)
+        checkRecord(records[3], setId, 1980, 30, "AGO", "cohort_size", 2000F)
+        checkRecord(records[4], setId, 1980, 30, "AGO", "dalys", 73.6F)
+        checkRecord(records[5], setId, 1980, 30, "AGO", "deaths", 20F)
     }
 
     protected fun checkModelRuns(db: JooqContext, modelRunData: ModelRunTestData)
@@ -190,7 +190,7 @@ abstract class BurdenEstimateRepositoryTests : RepositoryTests<BurdenEstimateRep
     }
 
     private fun checkRecord(record: Record, setId: Int,
-                            year: Short, age: Short, country: String, outcomeCode: String, outcomeValue: BigDecimal)
+                            year: Short, age: Short, country: String, outcomeCode: String, outcomeValue: Float)
     {
         val t = Tables.BURDEN_ESTIMATE
         Assertions.assertThat(record[t.BURDEN_ESTIMATE_SET]).isEqualTo(setId)
@@ -202,13 +202,13 @@ abstract class BurdenEstimateRepositoryTests : RepositoryTests<BurdenEstimateRep
     }
 
     protected fun data(runs: List<String?> = listOf(null, null)) = sequenceOf(
-            BurdenEstimateWithRunId(diseaseId, runs[0], 2000, 50, "AFG", "Afghanistan", 1000.toDecimal(), mapOf(
-                    "deaths" to 10.toDecimal(),
-                    "cases" to 100.toDecimal()
+            BurdenEstimateWithRunId(diseaseId, runs[0], 2000, 50, "AFG", "Afghanistan", 1000F, mapOf(
+                    "deaths" to 10F,
+                    "cases" to 100F
             )),
-            BurdenEstimateWithRunId(diseaseId, runs[1], 1980, 30, "AGO", "Angola", 2000.toDecimal(), mapOf(
-                    "deaths" to 20.toDecimal(),
-                    "dalys" to 73.6.toDecimal()
+            BurdenEstimateWithRunId(diseaseId, runs[1], 1980, 30, "AGO", "Angola", 2000F, mapOf(
+                    "deaths" to 20F,
+                    "dalys" to 73.6F
             ))
     )
 
@@ -230,12 +230,12 @@ abstract class BurdenEstimateRepositoryTests : RepositoryTests<BurdenEstimateRep
     protected fun checkStochasticBurdenEstimates(db: JooqContext, setId: Int)
     {
         val records = getStochasticEstimatesInOrder(db)
-        checkStochasticRecord(records[0], setId, 2000, 50, "AFG", "cases", 100.toDecimal())
-        checkStochasticRecord(records[1], setId, 2000, 50, "AFG", "cohort_size", 1000.toDecimal())
-        checkStochasticRecord(records[2], setId, 2000, 50, "AFG", "deaths", 10.toDecimal())
-        checkStochasticRecord(records[3], setId, 1980, 30, "AGO", "cohort_size", 2000.toDecimal())
-        checkStochasticRecord(records[4], setId, 1980, 30, "AGO", "dalys", 73.6.toDecimal())
-        checkStochasticRecord(records[5], setId, 1980, 30, "AGO", "deaths", 20.toDecimal())
+        checkStochasticRecord(records[0], setId, 2000, 50, "AFG", "cases", 100F)
+        checkStochasticRecord(records[1], setId, 2000, 50, "AFG", "cohort_size", 1000F)
+        checkStochasticRecord(records[2], setId, 2000, 50, "AFG", "deaths", 10F)
+        checkStochasticRecord(records[3], setId, 1980, 30, "AGO", "cohort_size", 2000F)
+        checkStochasticRecord(records[4], setId, 1980, 30, "AGO", "dalys", 73.6F)
+        checkStochasticRecord(records[5], setId, 1980, 30, "AGO", "deaths", 20F)
     }
 
     protected fun checkStochasticModelRuns(db: JooqContext, modelRunData: ModelRunTestData)
@@ -271,7 +271,7 @@ abstract class BurdenEstimateRepositoryTests : RepositoryTests<BurdenEstimateRep
     protected fun checkStochasticRecord(
             record: Record, setId: Int,
             year: Short, age: Short, country: String,
-            outcomeCode: String, outcomeValue: BigDecimal
+            outcomeCode: String, outcomeValue: Float
     )
     {
         val t = Tables.BURDEN_ESTIMATE_STOCHASTIC
